ml-service/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
requests
